Fixes two common javascript errors.
The first one is the fallback system of the translator subsystem incorrectly falls back to xpath. This is because older versions of the selenium IDE sometimes recorded the `css` type locators without the `css=` prefix, and nowadays it does the same with `xpath`.

The second one fixes an error which happens because it's assumed that a jQuery object's `tagName` property always returns an object.

Conflicts:
	compiled.js